Revert "pull from multientity branch"
This reverts commit f0190d8023901ca9835a2b1b61dea16572fcefa0, reversing
changes made to 3c367b4cc146c27bdc70a960515f6a9d93e7aa2e.

+++ src/main/java/in/handyman/raven/lib/model/multi/member/indicator/MultiValueMemberMapperInputTable.java
@@ -13,12 +13,12 @@ public class MultiValueMemberMapperInputTable {
 
     private Long rootPipelineId;
     private String batchId;
-    private Integer groupId;
+    private Long groupId;
     private String originId;
     private String sorContainerName;
     private String sorItemName;
     private String predictedValue;
-    private Integer paperNo;
+    private Long paperNo;
     private Long tenantId;
     private String encryptionPolicy;
     private String isEncrypted;
